requirements.txt
aioping
paho-mqtt
